tether_main/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.2
django-allauth==65.0.2
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-crontab==0.7.1
django-timezone-field==7.1
djangorestframework==3.15.2
kombu==5.5.3
prompt_toolkit==3.0.51
PyMySQL==1.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==6.0.0
six==1.17.0
sqlparse==0.5.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
